+++ scratch.js
@@ -5,23 +5,12 @@ const wtf = require('./src/index')
 // const wtf = require('./builds/wtf_wikipedia');
 // const wtf = require('./build');
 
-wtf.extend((models, templates) => {
-  templates.cool = (tmpl, r) => {
-    r.templates.push({ yes: true })
-    return 'hi'
-  }
-})
-let str = `
+// wtf.extend((models, templates) => {
+//   templates.cool = (tmpl, r) => {
+//     r.templates.push({ yes: true })
+//     return 'hi'
+//   }
+// })
 
-==One==
-asdfasdf
-
-==Two==
-also lkjasdlfkjdsf
-`
-
-let doc = wtf(`he is {{cool|true}}  {{fun|yep}}`)
-// let doc = wtf(`he is [[cool|true]]  [[fun|yep]]`)
-// let doc = wtf(str)
-console.log(doc.templates())
-// console.log(doc.templates())
+let doc = wtf(`hello [[world]]`)
+console.log(doc.links())