+++ src/link/Link.js
@@ -6,10 +6,25 @@ const Link = function(data) {
 }
 const methods = {
   text: function() {
-    return ''
+    return this.data.text
   },
   json: function() {
     return this.data
+  },
+  page: function() {
+    return this.data.page
+  },
+  anchor: function() {
+    return this.data.anchor
+  },
+  wiki: function() {
+    return this.data.wiki
+  },
+  site: function() {
+    return this.data.site
+  },
+  type: function() {
+    return this.data.type
   }
 }
 Object.keys(methods).forEach(k => {