+++ src/list/List.js
@@ -32,7 +32,7 @@ const methods = {
     } else if (typeof n === 'string') {
       //grab a link like .links('Fortnight')
       n = n.charAt(0).toUpperCase() + n.substring(1) //titlecase it
-      let link = links.find(o => o.page === n)
+      let link = links.find(o => o.page() === n)
       return link === undefined ? [] : [link]
     }
     return links