+++ src/reference/Reference.js
@@ -22,7 +22,7 @@ const methods = {
     } else if (typeof n === 'string') {
       //grab a link like .links('Fortnight')
       n = n.charAt(0).toUpperCase() + n.substring(1) //titlecase it
-      let link = arr.find(o => o.page === n)
+      let link = arr.find(o => o.page() === n)
       return link === undefined ? [] : [link]
     }
     return arr || []