+++ tests/images.test.js
@@ -24,7 +24,7 @@ The paintings have the freedom and energy of  sketches, using warm and cool ligh
   var templ = doc.templates(0)
   t.deepEqual(templ.template, 'gallery', 'document-has-template')
   t.deepEqual(templ.images.length, 5, '5 images')
-  t.deepEqual(templ.images[0].caption.links(0).page, 'Freyja', 'image has caption')
+  t.deepEqual(templ.images[0].caption.links(0).page(), 'Freyja', 'image has caption')
   t.deepEqual(doc.images().length, 5, 'images() finds gallery')
   t.end()
 })