+++ tests/misc.test.js
@@ -38,16 +38,16 @@ test('external links', t => {
   var str = `The [http://w110.bcn.cat/portal/site/Eixample] is the quarter designed`
   var obj = wtf(str)
   var link = obj.sentences(0).links(0)
-  t.equal(link.text, '', 'link-text')
-  t.equal(link.site, 'http://w110.bcn.cat/portal/site/Eixample', 'link-site')
-  t.equal(link.type, 'external', 'link-type')
+  t.equal(link.text(), '', 'link-text')
+  t.equal(link.site(), 'http://w110.bcn.cat/portal/site/Eixample', 'link-site')
+  t.equal(link.type(), 'external', 'link-type')
 
   str = `The [http://w110.bcn.cat/portal/site/Eixample Fun Times] is the quarter designed`
   obj = wtf(str)
   link = obj.sentences(0).links(0)
-  t.equal(link.text, 'Fun Times', 'link-text')
-  t.equal(link.site, 'http://w110.bcn.cat/portal/site/Eixample', 'link-site')
-  t.equal(link.type, 'external', 'link-type')
+  t.equal(link.text(), 'Fun Times', 'link-text')
+  t.equal(link.site(), 'http://w110.bcn.cat/portal/site/Eixample', 'link-site')
+  t.equal(link.type(), 'external', 'link-type')
   t.end()
 })